interpolation-related classes use _traits approach
Following the model chosen for Mesh, the interpolation classes now don't cast to
MeshInterface<meshType> but rather use meshType directly.
In case we need some intrinsics of the interpolation implementation, there's a
templated Interpolator_Impl_traits class for that. (I guess we'll need it for
compile-time check against MeshNodeType being the same for source/target
meshes).