More flexible generator scripts #254
Generation to c header files is disabled right now, since I don't want to deal with converting the
values to uint16_t and such.

Now I have a fairly good understanding of how SkinWare and encoding the values is supposed to work:
1. You take a "snapshot" of the sensor values in a no contact case
2. For subsequent readings of sensors, you subtract the initial "snapshot" and than use this result
   (which, mind you, need not be positive, su use int64_t or something similar) for the scaling
   algorithm.

It's all doable, but it's not a priority right now.